+++ crates/js-sys/src/lib.rs
@@ -9873,6 +9873,8 @@ impl UpcastFrom<JsString> for &str {}
 impl UpcastFrom<char> for JsString {}
 impl UpcastFrom<JsString> for char {}
 
+impl wasm_bindgen::__rt::marker::JsStringLikeSealed for JsString {}
+impl wasm_bindgen::__rt::marker::JsStringLikeSealed for &JsString {}
 impl wasm_bindgen::JsStringLike for JsString {}
 impl wasm_bindgen::JsStringLike for &JsString {}